elf: increase resolution of Atom.relocs_section_index to u32

+++ src/link/Elf/Atom.zig
@@ -20,7 +20,7 @@ input_section_index: u16 = 0,
 output_section_index: u16 = 0,
 
 /// Index of the input section containing this atom's relocs.
-relocs_section_index: u16 = 0,
+relocs_section_index: u32 = 0,
 
 /// Index of this atom in the linker's atoms table.
 atom_index: Index = 0,
